Move message component to component folder
